backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.0.0",
    "boto3>=1.28.0",
    "constructs>=10.0.0",
    "pillow>=9.0.0",
    "pytest==6.2.5",
]
